refactor(weather): centralize limitDecimals utility
Move limitDecimals function to shared utils.js to eliminate duplication between yr and smhi providers. Uses truncation instead of rounding to ensure coordinates stay within intended boundaries.
